requirements.txt
Flask==3.1.3
gunicorn==25.1.0
validators==0.35.0
python-whois==0.9.6
dnspython==2.8.0
email-validator==2.3.0
